pyproject.toml
[project]
name = "headliz"
version = "0.0.2"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gabliz", email = "lizzosgabriele@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "shellingham>=1.5.4",
    "typer[all]>=0.24.1",
    "uvicorn[standard]>=0.44.0",
]

[project.scripts]
headliz = "headliz.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["headliz"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
